+++ WeatherProjectTestOrion/Main/Model/ListDailyWeather.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import SwiftUI
+
 struct DayWeather: Hashable {
     let date: String
     let nameDay: String
@@ -18,20 +19,19 @@ struct DayWeather: Hashable {
 
 let listDaily = [
     DayWeather(date: "6 May", nameDay: "Today", icon: "cloud.rain.fill", colorIcon: .blue, temp1: "+10°", temp2: "+8°"),
-
-DayWeather(date: "7 May", nameDay: "Tomorrow", icon: "sun.max.fill", colorIcon: .yellow, temp1: "+21°", temp2: "+18°"),
-
-DayWeather(date: "8 May", nameDay: "Friday", icon: "cloud.rain.fill", colorIcon: .blue, temp1: "+19°", temp2: "+15°"),
-
-DayWeather(date: "9 May", nameDay: "Saturday", icon: "cloud.rain.fill", colorIcon: .blue, temp1: "+17°", temp2: "+13°"),
-
+    DayWeather(date: "7 May", nameDay: "Tomorrow", icon: "sun.max.fill", colorIcon: .yellow, temp1: "+21°", temp2: "+18°"),
+    DayWeather(date: "8 May", nameDay: "Friday", icon: "cloud.rain.fill", colorIcon: .blue, temp1: "+19°", temp2: "+15°"),
+    DayWeather(date: "9 May", nameDay: "Saturday", icon: "cloud.rain.fill", colorIcon: .blue, temp1: "+17°", temp2: "+13°"),
     DayWeather(date: "10 May", nameDay: "Sunday", icon: "sun.max.fill", colorIcon: .yellow, temp1: "+16°", temp2: "+14°"),
-
-DayWeather(date: "11 May", nameDay: "Monday", icon: "sun.max.fill", colorIcon: .yellow, temp1: "+21°", temp2: "+20°"),
-
-DayWeather(date: "12 May", nameDay: "Tuesday", icon: "cloud.rain.fill", colorIcon: .blue, temp1: "+19°", temp2: "+14°"),
-
-DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
-DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°")
-
+    DayWeather(date: "11 May", nameDay: "Monday", icon: "sun.max.fill", colorIcon: .yellow, temp1: "+21°", temp2: "+20°"),
+    DayWeather(date: "12 May", nameDay: "Tuesday", icon: "cloud.rain.fill", colorIcon: .blue, temp1: "+19°", temp2: "+14°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°"),
+    DayWeather(date: "13 May", nameDay: "Wednesday", icon: "cloud.bolt.fill", colorIcon: .blue, temp1: "+12°", temp2: "+8°")
 ]
